refactor : Router 구조 개선 및 home 화면에 Outlet 추가
