--- src/pages/Save.tsx
@@ -1,5 +1,0 @@
-const Save = () => {
-  return <div>Save</div>;
-};
-
-export default Save;
