+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import Home from './Home';
-import Save from './Save';
+import Save from './save/Save';
 import Search from './Search';
 import Example from './Example';
